+++ tests/test_checker/test_noqa.py
@@ -162,6 +162,7 @@ def test_noqa_fixture_disabled(absolute_path, all_violations):
         'Z465': 1,
         'Z466': 1,
         'Z467': 1,
+        'Z468': 1,
     }
 
     process = subprocess.Popen(