+++ tests/test_visitors/test_ast/test_compares/test_literal.py
@@ -3,15 +3,13 @@
 import pytest
 
 from wemake_python_styleguide.violations.consistency import (
-    ConstantComparisonViolation,
+    ConstantCompareViolation,
 )
-from wemake_python_styleguide.visitors.ast.comparisons import (
-    ComparisonSanityVisitor,
-)
+from wemake_python_styleguide.visitors.ast.compares import CompareSanityVisitor
 
-if_with_chained_comparisons1 = 'if 0 < {0} < {1}: ...'
-if_with_chained_comparisons2 = 'if {0} > {1} > 0: ...'
-if_with_chained_comparisons3 = 'if -1 > {0} > {1} > 0: ...'
+if_with_chained_compares1 = 'if 0 < {0} < {1}: ...'
+if_with_chained_compares2 = 'if {0} > {1} > 0: ...'
+if_with_chained_compares3 = 'if -1 > {0} > {1} > 0: ...'
 
 
 @pytest.mark.parametrize('comparators', [
@@ -26,10 +24,10 @@ def test_non_literal(
     comparators,
     default_options,
 ):
-    """Testing that comparisons work well."""
+    """Testing that compares work well."""
     tree = parse_ast_tree(simple_conditions.format(*comparators))
 
-    visitor = ComparisonSanityVisitor(default_options, tree=tree)
+    visitor = CompareSanityVisitor(default_options, tree=tree)
     visitor.run()
 
     assert_errors(visitor, [])
@@ -48,18 +46,18 @@ def test_literal(
     comparators,
     default_options,
 ):
-    """Testing that violations are when using literal comparisons."""
+    """Testing that violations are when using literal compares."""
     tree = parse_ast_tree(simple_conditions.format(*comparators))
 
-    visitor = ComparisonSanityVisitor(default_options, tree=tree)
+    visitor = CompareSanityVisitor(default_options, tree=tree)
     visitor.run()
 
-    assert_errors(visitor, [ConstantComparisonViolation])
+    assert_errors(visitor, [ConstantCompareViolation])
 
 
 @pytest.mark.parametrize('code', [
-    if_with_chained_comparisons1,
-    if_with_chained_comparisons3,
+    if_with_chained_compares1,
+    if_with_chained_compares3,
 ])
 @pytest.mark.parametrize('comparators', [
     (1, 'first_name'),
@@ -75,15 +73,15 @@ def test_literal_special1(
     """Testing that special cases do work and raise warnings."""
     tree = parse_ast_tree(code.format(*comparators))
 
-    visitor = ComparisonSanityVisitor(default_options, tree=tree)
+    visitor = CompareSanityVisitor(default_options, tree=tree)
     visitor.run()
 
-    assert_errors(visitor, [ConstantComparisonViolation])
+    assert_errors(visitor, [ConstantCompareViolation])
 
 
 @pytest.mark.parametrize('code', [
-    if_with_chained_comparisons2,
-    if_with_chained_comparisons3,
+    if_with_chained_compares2,
+    if_with_chained_compares3,
 ])
 @pytest.mark.parametrize('comparators', [
     ('first_name', 1),
@@ -99,16 +97,16 @@ def test_literal_special2(
     """Testing that special cases do work and raise warnings."""
     tree = parse_ast_tree(code.format(*comparators))
 
-    visitor = ComparisonSanityVisitor(default_options, tree=tree)
+    visitor = CompareSanityVisitor(default_options, tree=tree)
     visitor.run()
 
-    assert_errors(visitor, [ConstantComparisonViolation])
+    assert_errors(visitor, [ConstantCompareViolation])
 
 
 @pytest.mark.parametrize('code', [
-    if_with_chained_comparisons1,
-    if_with_chained_comparisons2,
-    if_with_chained_comparisons3,
+    if_with_chained_compares1,
+    if_with_chained_compares2,
+    if_with_chained_compares3,
 ])
 def test_literal_special_without_errors(
     assert_errors,
@@ -119,7 +117,7 @@ def test_literal_special_without_errors(
     """Testing that special cases do work and do not raise warnings."""
     tree = parse_ast_tree(code.format('first_name', 'second_name'))
 
-    visitor = ComparisonSanityVisitor(default_options, tree=tree)
+    visitor = CompareSanityVisitor(default_options, tree=tree)
     visitor.run()
 
     assert_errors(visitor, [])